# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payment Gateway API",
        "description": "API платёжного шлюза с интеграцией YooKassa",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/webhook/yookassa": {
            "post": {
                "tags": [
                    "Webhook"
                ],
                "summary": "Webhook от YooKassa",
                "description": "Принимает уведомления от YooKassa об изменении статуса платежа. Доступен только с IP-адресов YooKassa.",
                "operationId": "dbf01c97b31a23fc7293d41a58d6cef6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "event": {
                                        "type": "string",
                                        "example": "payment.succeeded",
                                        "enum": [
                                            "payment.succeeded",
                                            "payment.canceled",
                                            "refund.succeeded"
                                        ]
                                    },
                                    "object": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "22d65900-000f-5000-a000-10d000000000"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "succeeded"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook обработан"
                    },
                    "403": {
                        "description": "Запрос с неразрешённого IP"
                    }
                }
            }
        }
    },
...
}